# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
      "Board": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "projectId": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "createdAt"
        ],
        "type": "object"
      },
      "Column": {
        "properties": {
          "boardId": {
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "position": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "boardId",
          "name",
          "position",
          "createdAt"
        ],
        "type": "object"
      },
      "CreateBoardRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "projectId": {
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "projectId",
          "name"
        ],
        "type": "object"
      },
      "CreateColumnRequest": {
        "properties": {
          "boardId": {
            "minimum": 1,
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "position": {
            "type": "integer"
          }
        },
        "required": [
          "boardId",
          "name"
        ],
        "type": "object"
      },
      "CreatePrincipalRequest": {
        "properties": {
          "displayName": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "kind": {
            "enum": [
              "human",
              "agent"
            ],
            "type": "string"
          }
        },
        "required": [
          "kind",
          "handle",
          "displayName"
        ],
        "type": "object"
      },
      "CreateProjectRequest": {
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "PaginatedPrincipals": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Principal"
            },
            "type": "array"
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "totalItems": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "totalItems",
          "totalPages"
        ],
        "type": "object"
      },
...
      "Principal": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "kind": {
            "enum": [
              "human",
              "agent"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "handle",
          "displayName",
          "createdAt"
        ],
        "type": "object"
      },
      "Project": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "createdAt"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TODO App API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/boards": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Board"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Boards list"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "List boards"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Create board"
      }
    },
...
    "/api/columns": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "boardId",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Column"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Columns list"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "List columns"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateColumnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Column"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Create column"
      }
    },
...
    "/api/principals": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "schema": {
              "enum": [
                "human",
                "agent"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPrincipals"
                }
              }
            },
            "description": "Paginated principals"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "List principals"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrincipalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Principal"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Duplicate handle"
          }
        },
        "summary": "Create principal"
      }
    },
    "/api/projects": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Projects list"
          }
        },
        "summary": "List projects"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Create project"
      }
    },
    "/api/reviews/weekly": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "thresholdDays",
            "schema": {
              "default": 14,
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyReviewResponse"
                }
              }
            },
            "description": "Weekly review payload"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Weekly review sections"
      }
    },
    "/api/tasks": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "enum": [
                "inbox",
                "next",
                "waiting",
                "scheduled",
                "done",
                "someday",
                "reference"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "projectId",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "assigneeId",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "boardColumnId",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTasks"
                }
              }
            },
            "description": "Paginated tasks"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "List tasks"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Create task"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ]
}